backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google AI
google-generativeai==0.3.2
google-cloud-aiplatform==1.38.0

# LangChain
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.13

# Vector Database
chromadb==0.4.22

# Financial Data
yfinance==0.2.35
alpha-vantage==2.3.1
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# Utilities
python-dotenv==1.0.0
aiohttp==3.9.1

# Monitoring
prometheus-client==0.19.0